ntc: remove useless "fail" log messages
From log messages, there are lots of useless thermal ntc logs:
"user_vsensor_name write fail, 7"
It will compare the length of sensor name with that of variable
" USER_VSENSOR_NAME". But USER_VSENSOR_NAME is a macro
definition with the value of 32:
So the result of comparing will always fail.
It is useless for thermal ntc logic. So we will remove it.

Change-Id: I7344876e23a0759e4159e8ba42d92ed8d11a920b
Reviewed-on: https://gerrit.mot.com/2862838
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key