drm/mediatek: rollback bt2020 layer to GPU under native mode
Bilibili play hdr video, the output data is RGB888 +BT2020.
For MTK platform, RGB888 is recognized as UI, not video with YUV.
So MML_PQ will not handle the HDR data,So the data with BT020 wiil be output to
HWC, and there is not WCG for mt6878. So it can not hanble the data with BT2020.
So must Rollback o GPU, that GPU will transfer it.

Change-Id: I673c61c4a9779c3eac228149fbf74c670f07feeb
Signed-off-by: houdz1 <houdz1@lenovo.com>
Reviewed-on: https://gerrit.mot.com/3032953
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Zhihong Kang <kangzh@motorola.com>
Reviewed-by: Hua Tan <tanhua1@motorola.com>
Submit-Approved: Jira Key
Signed-off-by: tanhua1 <tanhua1@motorola.com>
Reviewed-on: https://gerrit.mot.com/3040112
Reviewed-by: Chuwei Xu <xucw@motorola.com>